# integrity-tech-frontend/src/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/evaluations/cat/sessions": {
      "post": {
        "operationId": "CatController_startSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCatSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sesión iniciada con éxito. Devuelve la sesión y el primer reactivo."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Iniciar una sesión adaptativa computerizada (CAT)",
        "tags": [
          "CAT (Computerized Adaptive Testing Engine)"
        ]
      }
    },
    "/api/evaluations/cat/sessions/{sessionId}/response": {
      "post": {
        "operationId": "CatController_submitResponse",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCatResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta procesada. Devuelve el siguiente reactivo o finalización."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar respuesta al reactivo adaptativo activo",
        "tags": [
          "CAT (Computerized Adaptive Testing Engine)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Integrity-Tech | Platform API",
    "description": "Especificación técnica de la API REST para el motor transaccional de evaluaciones psicométricas, telemetría de proctoring y feedback de PMF.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingresa tu Token JWT de sesión para autenticar las peticiones",
        "in": "header"
      }
    },
    "schemas": {
...
      "StartCatSessionDto": {
        "type": "object",
        "properties": {}
      },
      "SubmitCatResponseDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}